chore: Move rpc integration tests to unit tests, guaranteed to run sequentially thanks to nextest
